Organizing GUI task wise
